chore: separated webpack.config for dev and prod environment
