+++ package.json
@@ -32,13 +32,14 @@
     "style-loader": "^0.20.3",
     "webpack": "^4.1.1",
     "webpack-cli": "^2.0.11",
-    "webpack-dev-server": "^3.1.1"
+    "webpack-dev-server": "^3.1.1",
+    "webpack-merge": "^4.1.2"
   },
   "scripts": {
-    "build:prod": "./node_modules/.bin/webpack --display-entrypoints --mode production && gulp responsive-images-prod && npm run minify-html",
+    "build:prod": "./node_modules/.bin/webpack --display-entrypoints --mode production --config webpack.prod.js && gulp responsive-images-prod && npm run minify-html",
     "minify-html": "./node_modules/.bin/html-minifier --file-ext html --input-dir dist --output-dir dist --remove-comments --collapse-whitespace --minify-js --minify-css",
     "start:dev": "gulp responsive-images-dev && npm run start:dev-server",
-    "start:dev-server": "./node_modules/.bin/webpack-dev-server --devtool source-maps --hot --progress --colors --mode development",
+    "start:dev-server": "./node_modules/.bin/webpack-dev-server --devtool source-maps --hot --progress --colors --mode development --config webpack.dev.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [
