+++ src/js/GoogleMapsLoader.js
@@ -80,12 +80,9 @@ export default class GoogleMapsLoader {
     }
 
     load(options) {
-        // Object.entries(options).forEach(
-        //     ([key, value]) => this[key] = value
-        // );
-        for(let key of Object.keys(options)){
-            this[key] = options[key];
-        }
+        Object.entries(options).forEach(
+            ([key, value]) => this[key] = value
+        );
 
         let that = this;
         return new Promise(
